requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
langchain==0.0.335
langchain-community==0.0.13
langchain-core==0.1.4
langchain-google-genai==0.0.5
google-generativeai==0.3.1
pandas==2.1.3
chroma-hnswlib==0.7.3
chromdb==0.4.18
chromadb==0.4.18